fix : ResponseEntity를 이용한 기능 구현
- 기존 기능에 ResponseEntity를 이용
- 도메인과 구분하기 위한 DTO 구현
- 각 기능별 예외에 따른 코드 추가

* 기능구현
getAllProduct() : 상품 전체조회
getProduct(int) : 상품 상세조회
saveProduct(Product) : 상품 등록
updateProduct(int, ProductRequestDTO) : 상품 수정
deleteProduct(int) : 상품 삭제

* files
ProductController.java
ProductService.java
ProductRepository.java
ProductMapRepository.java
ProductRequestDTO.java
Product.java

+++ src/main/java/com/basecamp/ticketpick/product/ProductController.java
@@ -1,6 +1,10 @@
 package com.basecamp.ticketpick.product;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.Map;
 
 @RestController
 public class ProductController {
@@ -11,27 +15,55 @@ public class ProductController {
     }
 
     @GetMapping("/products")
-    public String getAllProduct() {
-        return productService.getAllProduct();
+    public ResponseEntity<String> getAllProduct() {
+        Map<Integer, Product> products = productService.getAllProduct();
+        if(products.isEmpty()){
+            return ResponseEntity.ok("no products.");
+        }
+        return ResponseEntity.ok(products.toString());
     }
 
     @GetMapping("/products/{id}")
-    public Product getProduct(@PathVariable int id) {
-        return productService.getProduct(id);
+    public ResponseEntity<?> getProduct(@PathVariable int id) {
+        Product product = productService.getProduct(id);
+        if(product==null){
+            return ResponseEntity.ok("There is not product, id="+id);
+        }
+        return ResponseEntity.ok(product);
     }
 
     @PostMapping("/products")
-    public String saveProduct(@RequestBody Product product) {
-        return productService.saveProduct(product);
+    public ResponseEntity<?> saveProduct(@RequestBody Product product) {
+        if(product==null){
+            return ResponseEntity.badRequest().body("Request data is null.");
+        }
+        int id = productService.saveProduct(product);
+        return ResponseEntity.created(URI.create("/products/"+id)).build();
     }
 
     @PatchMapping("/products/{id}")
-    public String updateProduct(@RequestBody Product product) {
-        return productService.updateProduct(product);
+    public ResponseEntity<?> updateProduct(@PathVariable int id, @RequestBody ProductRequestDTO productRequestDTO) {
+        Product product = productService.getProduct(id);
+        if(productRequestDTO==null){
+            return ResponseEntity.badRequest().body("Request data is null.");
+        }
+        if(product==null){
+            return ResponseEntity.badRequest().body("There is not product, id="+id);
+        }
+        productService.updateProduct(product, productRequestDTO);
+        return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/products/{id}")
-    public String deleteProduct(@PathVariable int id) {
-        return productService.deleteProduct(id);
+    public ResponseEntity<?> deleteProduct(@PathVariable int id) {
+        Product product = productService.getProduct(id);
+        if(product==null){
+            return ResponseEntity.ok("There is not product, id="+id);
+        }
+        if(id < 1){
+            return ResponseEntity.badRequest().body("Product id is under 1");
+        }
+        productService.deleteProduct(id);
+        return ResponseEntity.noContent().build();
     }
 }
